Split complication settings views, add floating live preview
- Split ComplicationListView.swift into one view per file
  (ComplicationsRootView, ComplicationListView,
  WatchComplicationBuilderEditView), per the one-struct-per-file rule;
  the legacy list gained its missing #Preview. Surface that rule in a
  new Non-Negotiable Conventions section of AGENTS.md.
- Make FakeServerManager public and seed it in the builder #Preview so
  the server-gated form (including the complication mock) renders in
  Xcode previews.
- New FloatingPanel component in HADesignSystem: drag anywhere with
  corner snapping (flick-aware), tap or pinch to resize, configurable
  corner radius; registered in the components gallery.
- In the complication builder, the preview floats over the form once
  its inline section scrolls away, showing only the selected family
  centered and zoomed to fit a compact watch face whose bezel radius
  the panel chrome matches concentrically.
- Persist the editor's Customize disclosure as a new nullable
  isCustomized column (auto-added by the column migration); older rows
  fall back to inferring it from per-size overrides.
