+++ Sources/App/Settings/AppleWatch/Complications/Preview/AllFamiliesComplicationPreview.swift
@@ -4,10 +4,18 @@ import SwiftUI
 /// Shows the complication in all four WidgetKit families at once, arranged like an Apple Watch face,
 /// so the user sees every size simultaneously instead of flipping a size picker. Fetches the entity
 /// state once and renders every family from that single fetch.
+///
+/// With `showsOnlySelectedFamily` it instead renders just the selected family, centered and zoomed
+/// to fit a compact fake watch screen — the shape used by the floating (picture-in-picture) preview.
 struct AllFamiliesComplicationPreview: View {
     let config: WatchComplicationConfig
     let server: Server
     @Binding var selectedFamily: WatchComplicationConfig.Family
+    var showsOnlySelectedFamily = false
+
+    /// Outer bezel radius of the compact single-family face — exposed so the floating panel's
+    /// chrome can use a concentric radius that matches the fake watch shape.
+    static let compactBezelCornerRadius: CGFloat = 46
     /// Reports the entity's unit (nil when none), so the editor can offer the "Show unit" toggle.
     var onUnit: (String?) -> Void = { _ in }
     /// Reports the entity's attribute names (sorted), offered as value sources.
@@ -20,6 +28,9 @@ struct AllFamiliesComplicationPreview: View {
     @State private var entityAttributes: [String: Any] = [:]
     @State private var isFetching = false
     @State private var lastFetchKey: String?
+    /// Natural size of the selected family's preview, measured to compute the zoom-to-fit scale in
+    /// the single-family (floating) mode.
+    @State private var selectedPreviewSize: CGSize = .zero
 
     // Template rendering, used only for the custom-template kind.
     @StateObject private var valueRenderer: TemplateRenderer
@@ -29,6 +40,7 @@ struct AllFamiliesComplicationPreview: View {
         config: WatchComplicationConfig,
         server: Server,
         selectedFamily: Binding<WatchComplicationConfig.Family>,
+        showsOnlySelectedFamily: Bool = false,
         onUnit: @escaping (String?) -> Void = { _ in },
         onAttributes: @escaping ([String]) -> Void = { _ in },
         onValueIsNumeric: @escaping (Bool) -> Void = { _ in }
@@ -36,6 +48,7 @@ struct AllFamiliesComplicationPreview: View {
         self.config = config
         self.server = server
         self._selectedFamily = selectedFamily
+        self.showsOnlySelectedFamily = showsOnlySelectedFamily
         self.onUnit = onUnit
         self.onAttributes = onAttributes
         self.onValueIsNumeric = onValueIsNumeric
@@ -60,52 +73,99 @@ struct AllFamiliesComplicationPreview: View {
     }
 
     var body: some View {
-        ZStack(alignment: .topTrailing) {
-            ZStack {
-                RoundedRectangle(cornerRadius: 72, style: .continuous)
-                    .fill(
-                        LinearGradient(
-                            colors: [Color(white: 0.28), Color(white: 0.1)],
-                            startPoint: .topLeading,
-                            endPoint: .bottomTrailing
+        Group {
+            if showsOnlySelectedFamily {
+                // Compact single-family face for the floating preview: the current family only,
+                // centered and zoomed to fit the screen.
+                let screenSize = CGSize(width: 140, height: 176)
+                let available = CGSize(width: screenSize.width - 24, height: screenSize.height - 24)
+                // Capped so tiny content (e.g. the inline capsule) isn't blown up absurdly.
+                let fitScale: CGFloat = selectedPreviewSize == .zero ? 1 : min(
+                    1.5,
+                    available.width / max(selectedPreviewSize.width, 1),
+                    available.height / max(selectedPreviewSize.height, 1)
+                )
+                ZStack {
+                    RoundedRectangle(cornerRadius: Self.compactBezelCornerRadius, style: .continuous)
+                        .fill(
+                            LinearGradient(
+                                colors: [Color(white: 0.28), Color(white: 0.1)],
+                                startPoint: .topLeading,
+                                endPoint: .bottomTrailing
+                            )
                         )
-                    )
-                    .frame(width: 244, height: 300)
-                    .shadow(color: .black.opacity(0.25), radius: 14, y: 8)
+                        .frame(width: screenSize.width + 16, height: screenSize.height + 16)
 
-                RoundedRectangle(cornerRadius: 64, style: .continuous)
-                    .stroke(Color.white.opacity(0.18), lineWidth: 1)
-                    .frame(width: 230, height: 286)
-
-                RoundedRectangle(cornerRadius: 58, style: .continuous)
-                    .fill(Color.black)
-                    .frame(width: 214, height: 270)
-                    .overlay {
-                        ZStack {
-                            familyButton(.circular)
-                                .position(x: 58, y: 68)
-                            familyButton(.corner)
-                                .position(x: 156, y: 68)
-                            familyButton(.rectangular)
-                                .position(x: 107, y: 148)
-                            familyButton(.inline)
-                                .position(x: 107, y: 226)
+                    RoundedRectangle(cornerRadius: 40, style: .continuous)
+                        .fill(Color.black)
+                        .frame(width: screenSize.width, height: screenSize.height)
+                        .overlay {
+                            preview(for: selectedFamily)
+                                // Ideal size, so flexible content (the inline capsule) isn't squashed
+                                // by the small screen proposal before it gets scaled to fit.
+                                .fixedSize()
+                                .background(
+                                    GeometryReader { proxy in
+                                        Color.clear
+                                            .onAppear { selectedPreviewSize = proxy.size }
+                                            .onChange(of: proxy.size) { newSize in
+                                                selectedPreviewSize = newSize
+                                            }
+                                    }
+                                )
+                                .scaleEffect(fitScale)
                         }
-                        .frame(width: 214, height: 270)
-                        .clipShape(RoundedRectangle(cornerRadius: 58, style: .continuous))
-                    }
-            }
-            .frame(width: 260, height: 316)
+                        .clipShape(RoundedRectangle(cornerRadius: 40, style: .continuous))
+                }
+            } else {
+                ZStack(alignment: .topTrailing) {
+                    ZStack {
+                        RoundedRectangle(cornerRadius: 72, style: .continuous)
+                            .fill(
+                                LinearGradient(
+                                    colors: [Color(white: 0.28), Color(white: 0.1)],
+                                    startPoint: .topLeading,
+                                    endPoint: .bottomTrailing
+                                )
+                            )
+                            .frame(width: 244, height: 300)
+                            .shadow(color: .black.opacity(0.25), radius: 14, y: 8)
 
-            if isFetching {
-                ProgressView()
-                    .controlSize(.small)
-                    .tint(.white)
-                    .padding(DesignSystem.Spaces.two)
+                        RoundedRectangle(cornerRadius: 64, style: .continuous)
+                            .stroke(Color.white.opacity(0.18), lineWidth: 1)
+                            .frame(width: 230, height: 286)
+
+                        RoundedRectangle(cornerRadius: 58, style: .continuous)
+                            .fill(Color.black)
+                            .frame(width: 214, height: 270)
+                            .overlay {
+                                ZStack {
+                                    familyButton(.circular)
+                                        .position(x: 58, y: 68)
+                                    familyButton(.corner)
+                                        .position(x: 156, y: 68)
+                                    familyButton(.rectangular)
+                                        .position(x: 107, y: 148)
+                                    familyButton(.inline)
+                                        .position(x: 107, y: 226)
+                                }
+                                .frame(width: 214, height: 270)
+                                .clipShape(RoundedRectangle(cornerRadius: 58, style: .continuous))
+                            }
+                    }
+                    .frame(width: 260, height: 316)
+
+                    if isFetching {
+                        ProgressView()
+                            .controlSize(.small)
+                            .tint(.white)
+                            .padding(DesignSystem.Spaces.two)
+                    }
+                }
+                .frame(maxWidth: .infinity)
+                .padding(.vertical, DesignSystem.Spaces.two)
             }
         }
-        .frame(maxWidth: .infinity)
-        .padding(.vertical, DesignSystem.Spaces.two)
         .environment(\.colorScheme, .dark)
         // Re-run the fetch/render whenever a fetch input changes (entity, server, kind, template) —
         // reliably, so the preview updates on entity change without needing to tap a family first.
@@ -300,5 +360,19 @@ struct AllFamiliesComplicationPreview: View {
         }
         .padding()
     }
+}
+
+#Preview("Floating (single family)") {
+    LazyVGrid(columns: [GridItem(.flexible()), GridItem(.flexible())], spacing: DesignSystem.Spaces.three) {
+        ForEach(WatchComplicationConfig.Family.allCases) { family in
+            AllFamiliesComplicationPreview(
+                config: WatchComplicationConfig(serverId: "preview", name: "Solar"),
+                server: ServerFixture.standard,
+                selectedFamily: .constant(family),
+                showsOnlySelectedFamily: true
+            )
+        }
+    }
+    .padding()
 }
 #endif
